Calculs d'ampere et de voltages
Dans les classes des composantes… il va rester l’analyseur à faire…

+++ ProjetIntegrateur/src/projetintegrateur/Composante.java
@@ -4,6 +4,6 @@ public interface Composante {
     public int getNumero();
     public Type getType();
     public double getResistanceEquivalente();
-    public void calculAmpere(int voltage);
-    public void calculVoltage(int ampere);
+    public void calculAmpere(double voltage);
+    public void calculVoltage(double ampere);
 }
